Add feature to print each new ideal route
when working with the melbourne tram data set there are billions, or maybe trillions of solutions. If you run the code as is it will take weeks or years to finish.

Code updated to print out the route each time it find a new shorted route.

This is mostly just adding print statements but also needed a little refactor so that it can calculate each route as it is generated instead of generating a huge lists of routes and then returning them all.